REAL: arithmetic and comparison operator rework: enable an overload only if argument types are compatible
This prevents the situation that an arithmetical operator for REALs is invoked
on two arguments that are *convertible* to REAL but both of which are not REALs
already.

This is done by the new template enable_if_compat<Base,EquivBase,Compat>;
ensuring that at least one operand already is of type REAL.

Additionally, the assignment operators +=, -=, *= and /= are members of class
REAL now.